Make static pages loadable
