+++ src/common/constants/faq-keys-general.ts
@@ -1,0 +1,31 @@
+export const faqKeysGeneral = [
+  'what-is-ecency',
+  'what-is-difference',
+  'what-is-hive',
+  'why-choose-ecency',
+  'how-ecency-works',
+  'how-to-join',
+  'can-change-username',
+  'can-delete-account',
+  'how-to-signin',
+  'how-to-contribute',
+  'how-referrals-work',
+  'what-is-points',
+  'where-tokens-come',
+  'what-are-rc',
+  'how-promotion-work',
+  'how-boosting-work',
+  'why-refund-points',
+  'boosting-vote-buying',
+  'different-reward-choices',
+  'can-guest-post',
+  'can-link-post',
+  'source-label',
+  'how-to-transfer',
+  'how-see-rewards',
+  'when-claim-rewards',
+  'how-exchange-work',
+  'how-boost-account',
+  'how-dhf-work',
+  'what-spam-abuse',
+];
